# tauri/packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Springtale management API",
    "description": "Local-first bot automation. Binds loopback; bearer-authenticated.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/connectors/install": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "POST /connectors/install — install a connector from manifest JSON.",
        "operationId": "connectors_install",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connector installed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/install-wasm": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "POST /connectors/install-wasm — multipart with a `manifest` part\n(JSON text) and a `wasm` part (binary). Signature, hash and\ncapability checks all live in the operation.",
        "operationId": "connectors_install_wasm",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "WASM connector installed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/connectors/setup": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "POST /connectors/setup — configure and load an available connector.",
        "operationId": "connectors_setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connector set up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{name}": {
      "delete": {
        "tags": [
          "connectors"
        ],
        "summary": "DELETE /connectors/{name} — remove a connector.",
        "operationId": "connectors_remove",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Connector name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connector removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{name}/cascade": {
      "delete": {
        "tags": [
          "connectors"
        ],
        "summary": "DELETE /connectors/{name}/cascade — remove connector + dependent rules.",
        "operationId": "connectors_remove_cascade",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Connector name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connector and dependent rules removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{name}/config": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "GET /connectors/{name}/config — get connector config.",
        "operationId": "connectors_get_config",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Connector name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connector config (secrets redacted)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{name}/disable": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "POST /connectors/{name}/disable — disable a connector.",
        "operationId": "connectors_disable",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Connector name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connector disabled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{name}/enable": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "POST /connectors/{name}/enable — enable a connector.",
        "operationId": "connectors_enable",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Connector name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connector enabled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{name}/outputs": {
      "get": {
        "tags": [
          "connectors"
        ],
        "summary": "GET /connectors/{name}/outputs — list recent execution results.",
        "operationId": "connectors_list_outputs",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Connector name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Declared connector outputs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/connectors/{name}/reload": {
      "post": {
        "tags": [
          "connectors"
        ],
        "summary": "POST /connectors/{name}/reload — G4 hot-reload a connector without\ndropping in-flight calls. The connector's persisted config is\nre-applied verbatim; subsequent dispatches land on the new host.",
        "operationId": "connectors_reload",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Connector name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connector reloaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/ui": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Serve the dashboard index page (root of /ui/).",
        "operationId": "dashboard_serve_dashboard_index",
        "responses": {
          "200": {
            "description": "Dashboard SPA index",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {}
        ]
      }
    },
    "/ui/{path}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Serve an embedded dashboard file by path.",
        "description": "Falls back to index.html for SPA client-side routing.\nMount this at `/ui/` in the router.",
        "operationId": "dashboard_serve_dashboard",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "description": "Asset path under the SPA root",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard asset",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}